chore(astro): Make hotloading the default
