+++ packages/astro/bundled/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "../dist/bundled.js"
+}
