+++ packages/astro/src/bundled.ts
@@ -1,0 +1,1 @@
+export { default } from './integration/bundled';
