--- packages/astro/src/hotload.ts
@@ -1,1 +1,0 @@
-export { default } from './integration/hotload';
