+++ packages/astro/src/integration/bundled.ts
@@ -1,5 +1,5 @@
 import { createIntegration } from './create-integration';
 
 export default createIntegration({
-  mode: 'hotload',
+  mode: 'bundled',
 });
