+++ packages/astro/tsup.config.ts
@@ -8,16 +8,16 @@ export default defineConfig(() => {
     clean: true,
     entry: [
       './src/index.ts',
-      './src/hotload.ts',
+      './src/bundled.ts',
       './src/client/react/index.ts',
       './src/client/stores/index.ts',
       './src/client/index.ts',
-      './src/client/hotload.ts',
+      './src/client/bundled.ts',
       './src/server/index.ts',
       './src/internal/index.ts',
-      './src/internal/hotload.ts',
+      './src/internal/bundled.ts',
       './src/integration/index.ts',
-      './src/integration/hotload.ts',
+      './src/integration/bundled.ts',
       './src/async-local-storage.client.ts',
       './src/async-local-storage.server.ts',
     ],
